requirements.txt
streamlit==1.45.1
pandas==2.3.0
requests==2.32.4
python-dotenv==1.0.1
openpyxl==3.1.2
